Fix decoding timezone
mgo has an inconsistent behaviour with times indeed (see this: #go-mgo/mgo#13 and this: go-mgo/mgo#362). A UTC time value which is stored as UTC in Mongo, when you fetch it, mgo decodes it to the local time. I fixed it with returning the time in UTC. So, anyone can after convert it to whatever timezone they want.