Revert previous "Fix".
Turns out M is defined in another file, but because of an incorrect testing configuration on my side, the syntax checker
wasn't picking it up.  After moving some stuff around, it all works.

A couple of tests fail but I'm pretty confident that this is because I'm testing on a mac.