(wip) Remove on a document can pull the queue, too.
We are directly inserting the Queue, but instead of just blindly copying
what was in the old document, apply the $pullAll that we would have
applied if this were an Update. Because we are creating the stash
document from scratch, we don't need to ask Mongo to do the work.