Proper fix, including on Insert.
It turns out that I missed we had just reread the document as part of
the 'add txn-insert' or 'add txn-remove' to the document. That gives us
an updated txn-queue that needs to be handled. Since we are now cleaning
directly on the txn-queue instead of the cached queue docs, go ahead and
do it on both Insert and Remove.