Enable ReplicaSet in dbtest server.
Some parts of the mgo API (e.g. Change Streams) only work on
mongodb ReplicaSets. To enable testing for those, this commit
adds the ability to start a 1-member ReplicaSet via the test
server.
Enable it by setting the ReplicaSet boolean field on the DBServer
struct to true, before creating a Session.